% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{sunset_palettes}
\alias{sunset_palettes}
\alias{sunset_palettes_discrete}
\alias{sunset_palettes_sequential}
\title{Sunset palettes}
\format{
\subsection{\code{sunset_palettes}}{

An object of class \code{palettes_palette} with 7
colour palettes. Use \code{names(sunset_palettes)} to return all palette names.
}

\subsection{\code{sunset_palettes_discrete}}{

An object of class \code{palettes_palette} with
2 discrete colour palettes.
Use \code{names(sunset_palettes_discrete)} to return all palette names.
}

\subsection{\code{sunset_palettes_sequential}}{

An object of class \code{palettes_palette} with
5 sequential colour palettes.
Use \code{names(sunset_palettes_sequential)} to return all palette names.
}
}
\source{
\url{https://github.com/mccarthy-m-g/sunsets}
}
\usage{
sunset_palettes

sunset_palettes_discrete

sunset_palettes_sequential
}
\description{
Palettes inspired by sunsets in the Canadian Prairies.
}
\examples{
# Get all palettes by name.
names(sunset_palettes)

# Plot all palettes.
plot(sunset_palettes)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}, \code{\link[=pal_colour]{pal_colour()}}
}
\author{
\href{https://github.com/mccarthy-m-g}{Michael McCarthy}
}
\keyword{datasets}
