% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunsets-package.R
\docType{package}
\name{sunsets-package}
\alias{sunsets}
\alias{sunsets-package}
\title{sunsets: Sunset Colour Palettes}
\description{
Colour palettes inspired by sunsets in the Canadian Prairies. Built using the 'palettes' package, which provides methods for printing, formatting, casting and coercion, extraction and updating of components, plotting, colour mixing arithmetic, and colour interpolation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mccarthy-m-g/sunsets}
  \item Report bugs at \url{https://github.com/mccarthy-m-g/sunsets/issues}
}

}
\author{
\strong{Maintainer}: Michael McCarthy \email{m.mccarthy1624@gmail.com}

}
\keyword{internal}
